@@ -13,7 +13,7 @@
     }
   ],
   "require": {
-    "php": ">=5.4.0",
+    "php": ">=7.0",
     "yiisoft/yii2": "~2.0.13",
     "yiisoft/yii2-httpclient": "^2.0"
   },
